pyproject.toml
[tool.poetry]
name = "file-automation"
version = "0.1.0"
description = "File Automation - run templated shell commands for each file matched by pattern."
authors = ["Alexander Gorokhov <sashgorokhov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "<4"  # to support py3.8

[tool.poetry.scripts]
file_automation = 'file_automation.main:entrypoint'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
atomic = true
line_length = 120
lines_after_imports = 2
combine_as_imports = true

[tool.black]
line-length = 120
